--- examples/Google.Cloud.Functions.Examples.TimeZoneConverter/Startup.cs
@@ -1,33 +1,0 @@
-﻿// Copyright 2020, Google LLC
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     https://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-using Google.Cloud.Functions.Hosting;
-using Microsoft.AspNetCore.Hosting;
-using Microsoft.Extensions.DependencyInjection;
-using NodaTime;
-
-[assembly: FunctionsStartup(typeof(Google.Cloud.Functions.Examples.TimeZoneConverter.Startup))]
-
-namespace Google.Cloud.Functions.Examples.TimeZoneConverter
-{
-    /// <summary>
-    /// Startup class to provide the NodaTime built-in TZDB (IANA) date time zone provider
-    /// as a dependency to <see cref="Function"/>.
-    /// </summary>
-    public class Startup : FunctionsStartup
-    {
-        public override void ConfigureServices(WebHostBuilderContext context, IServiceCollection services) =>
-            services.AddSingleton(DateTimeZoneProviders.Tzdb);
-    }
-}
